made home ui into reusable container component
