+++ components/home/ui/HomeUI.tsx
@@ -1,37 +1,23 @@
 import React from "react";
-import Image from "next/image";
 
 import {
   useTheme,
-  Grid,
   Stack,
   Typography,
-  useMediaQuery,
   List,
   ListItemButton,
   ListItem,
-  Box,
 } from "@mui/material";
 import { blue } from "@mui/material/colors";
 
+import SectionImageRight from "../../section-containers/section-image-right";
+
 export default function HomeUI() {
   const theme = useTheme();
-  const matchesSmallerScreen = useMediaQuery(theme.breakpoints.down("md"));
 
   return (
-    <Grid
-      container
-      direction={matchesSmallerScreen ? "column-reverse" : "row"}
-      justifyContent="start"
-      sx={{ height: "100vh", pt: 9 }}
-    >
-      <Grid
-        p={8}
-        container
-        alignItems="center"
-        justifyContent="flex-end"
-        sx={{ width: { xs: "100%", md: "50%" } }}
-      >
+    <SectionImageRight
+      textSection={
         <Stack
           sx={{
             maxWidth: { xs: "100%", lg: "80%" },
@@ -63,33 +49,10 @@ export default function HomeUI() {
             {techStackButton("QA Testers")}
           </List>
         </Stack>
-      </Grid>
-      <Grid
-        sx={{
-          zIndex: -1,
-          width: { xs: "100%", md: "50%" },
-          height: { xs: "40vh", md: "100vh" },
-          position: "relative",
-        }}
-      >
-        <Box
-          sx={{
-            width: "100%",
-            height: "100%",
-            position: { xs: "relative", md: "static" },
-          }}
-        >
-          <Image
-            alt={"Login Image"}
-            objectFit={"cover"}
-            layout="fill"
-            src={
-              "https://images.unsplash.com/photo-1527689368864-3a821dbccc34?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=1350&q=80"
-            }
-          />
-        </Box>
-      </Grid>
-    </Grid>
+      }
+      imageAlt="Login Image"
+      imageSrc="https://images.unsplash.com/photo-1527689368864-3a821dbccc34?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=1350&q=80"
+    />
   );
 
   function techStackButton(text: string) {
